feat: adds .github/CODEOWNERS (#3992)
This file contains the usernames of the code's owners. The file syntax
is similar to that of .gitignore. The owners will be automatically added
to the review line of PRs. This should make it easier for us to request
PR reviews, and it will definitely be easier for OSS contributors who
want to send a PR.

I've not used this feature personally, but I think it's worth trying
out. We can always remove this if it's harmful or not helpful.