feat: remove tests from install Docker images (#4016)
This PR cleans up our docker "install" images a bit, by not installing
nor running the unit tests on these builds. In this project, "install"
builds intended to verify if our libraries are installed on a target
system, users can correctly link with them. googletest/benchmark is not
needed for this.