cleanup: refactor version() and version_string() (#4062)
We had multiple implementations of the version() and version_string()
functions, one on each library. This refactors the implementations to
use only the one from `google::cloud::`. I do not think we should remove
these functions, while rarely used they are part of the public API, and
we should avoid breaking changes when we can.

In the case of bigquery/ and pubsub/ we probably can remove the
functions, and I propose doing so in a followup PR.