ci: speed up clang-tidy builds (#4068)
This change minimizes the time spent on clang-tidy for PR builds.

On interactive builds and continuous builds the `clang-tidy` build works
as before: CMake calls `clang-tidy` for whatever TU it rebuilds, and so
(for the most part) only changed files run `clang-tidy`. Changes in
headers, of course, trigger a lot more `clang-tidy` work, but that might
be necessary.

On PR builds only the files changed w.r.t. to the target branch run through
`clang-tidy`. In some cases that might cause more files that necessary to
be examined, e.g. if the PR was based on an older version of `master`.
And in some it will cause fewer files, notably with header changes that
could affect the `clang-tidy` results.

The build times are about 8m with a warm cache.