ci: use emulator for the spanner integration tests (#4077)
For CMake builds we use the emulator to run the spanner integration
tests. The main motivation is the `coverage` test, which needs to
serialize the tests and runs very slowly against production. Other
builds run fine against the emulator or against production, but testing
locally is usually faster.

The drop in code coverage (0.22%) can be fixed with via #4080, but I think
we should leave that to a future PR.