feat: enable angle-bracket includes of Abseil headers
This creates an overlay BUILD file for the Abseil repo, which gives us a
target that will find Abseil's headers when included with
angle-brackets.

What are the downsides to doing this? I'm not 100% certain, but I think
one downside is that our bazel-using customers will also need to provide
this overlay BUILD target for Abseil. This is partly related to the way
bazel requires all users to list/fetch the dependencies of their
dependencies. However, we've already decided that it's OK to do this for
our googleapis.BUILD file, which would also force a requirement on our
bazel customers.

If we're OK doing it for one, why not both? Or maybe we should drop this
target from both and use double-quote includes?