fix: installed cmake files list absl as a dep (#4110)
Our installed cmake files were previously not listing absl as a dep, so callers were not able to link with our installed libraries and cmake files. This PR adds the correct cmake rules and also adds a non-test dep on absl::make_unique to common, bigtable, storage, and spanner, to show that the installed code actually works and links when running our quickstart tests.