ci: use emulator for Bigtable+Bazel integration tests (#4121)
We are running out of quota in production to run the Bigtable
integration tests. This change moves the Bazel-based builds to use the
emulators. A future PR will restore the nightly build to use production,
but we need to unblock progress first. The CMake-based builds already
use the emulator(s) for integration tests.