fix: warning options exported in library interface (#4134)
We mistakenly exported the options with the library, forcing all
applications that want to use the library to get the same options.

Historically we preferred creating an `INTERFACE` library, say
`storage_common_options` that would get all -Wall/-Werror options and we
would link this library *privately* against each target. That would add
these options to the target, but not its dependencies.

Because we had many of these '*_common_options' libraries we created a
function to add the `-Wall/-Werrror` options to them, and because it was
an interface library the options needed to be in the `INTERFACE`
section.

But *then* we started using this function to add the options to programs
and libraries directly, and that is where we went astray.

I am telling this story mostly to remind myself how we got here, and
maybe there is a lesson about re-purposing functions in there.