fix: reduce flakiness in macos/cmake-super build (#4148)
These builds are failing with "Socket closed" errors (reported as
exceptions by the sample code, but that is not important). Whatever the
underlying gRPC error is, the mutations changed in this PR were not
idempotent and therefore not retried by the library, and because they
are only sample data for the code, this was entirely unnecessary.

I expect that with these changes the errors will be less common, and the
builds will be less flaky, but I cannot say if they will go away.