fix: MSVC+CMake builds with GCS+gRPC disabled
With MSVC+CMake when we disable the GCS+gRPC plugin we cannot even
include the gRPC-related headers. This is because Protobuf requires some
magic `-D...` options in the command-line, and these are only present if
you link the Protobuf library, and the point of disabling the GCS+gRPC
plugin is to **not** link said library.

This goes undetected on Bazel, where we always build the GCS+gRPC plugin.