cleanup: move example_driver to testing_util/ (#4223)
Fixes: #4151

Moved the example_driver.* code from google/cloud/internal -> google/cloud/testing_util/. This makes it clearer that it's intended only for our "tests", and not for production code. This also avoids th need to compile and install the example driver.

I also changed bigtable and storage "examples" to only compile if (BUILD_TESTING). This avoids the need to compile them when a user is trying to compile a release build and install it.

I also cleaned up the old GOOGLE_CLOUD_CPP_TESTING_UTIL_ENABLE_INSTALL option, which is no longer needed since -common merged into the monorepo.