feat: add absl crash handler support for bigtable examples (#4150)
Part of #3990
Might help diagnose #4119

This PR adds a crash handler function that must be called early in main() to be used. This function will print a stack trace for any segfault, etc, and uncaught stack trace. This means that caught stack traces will not result in stack traces. I instrumented all of bigtable's example's main functions to install this crash handler.

Note: I also needed to disable the clang-tidy naming check for macros, because this code relies on a clang macro name that violates our style guide.