cleanup: make the test labels/tags more consistent (#4257)
Fixes #3868

This PR attempts to make our use of labels (cmake) and tags (bazel) a bit simpler and more consistent. This might be a small step in that direction:

Use the singular term "integration-test" everywhere
Removed all $product-integration-test labels. These are no longer needed. Products are identified in ctest invocations using -R $product, and in bazel invocations using bazel test google/cloud/$product/....
All integration tests in bazel and cmake have an integration-test label. This is most helpful when excluding all integration tests from a test run.
[CMAKE ONLY] All integration tests also have a label of either integration-test-emulator or integration-test-production, to identify whether or not they should run w/ an emulator. (there's one documented exception here about a bigquery integration test that just doesn't fully work today because it needs a script to invoke it w/ some arguments, and that's more than we need to do today).
That last bullet point suggests something odd: The bazel and cmake labels are not in sync. In bazel we have no indication of whether integration tests use an emulator or not. Instead, when we run integration tests in bazel, we manually pick and choose which tests do what (see build-in-docker-bazel.sh). It would be nice to clean that up in another PR.