fix: fix off-by-one in uploading streams to GCS (#4250)
Before this change, `seekg` was called for proper behavior for every
`UploadChink`. Because it was `expected`, which was broken, the `seekg`
was effectively a noop, so everything worked, except for generating
warnings.

This PR also removes the `seekg` from that execution path altogether.
Playing with tests proved that unless there is a bug in the code, there
is no way to end up in a situation when `seekg` was called and even if
this situation occurred, it would be impossible to properly recover from
it. In order to not increase the likelihood of a corruption, I replaced
this logic with returning an error.

In order to make sure that resumable upload sessions are properly
handled, a `seekg` is added at the beginning to handle resuming a
stream.