feat(storage/benchmark): reduce random data size (#4287)
We only upload one chunk at a time, and therefore we only need to
compute random data up to the maximum chunk size. We were computing
random data up to a full object size, which can be several times larger.

Also reduce the amount of entropy consumed to initialize the random
number generators.