feat: optionally use host network in docker containers (#4288)
When enabled with the DOCKER_USE_HOST_NETWORK environment variable, our docker containers will share the host's networking. This can be useful when running on a host with a non-trivial networking configuration, such as DNS caches, etc, that you want to share with the docker container.

This change also exposed an issue where some of our gcloud commands used
incidental gcloud accounts that existed on the system, other than the
service acount that we intended. This is fixed by passing --account=""
to some gcloud commands where we know we don't want to use any existing
system account.

I believe this should fix the issue that @devbww and I were having when running the integration tests on our GCE instances.