fix: enable spanner slow integration tests  (#4295)
These "slow" integration tests are only enabled for certain builds at
this point. For now, only the "integration" build (both CI and
presubmit), and builds that use the Spanner emulator.

In the previous repo, google-cloud-cpp-spanner, these tests were
controlled by the variable RUN_SLOW_INTEGRATION_TESTS, but in this
monorepo, that variable was too generic, so it was renamed to be
Spanner specific.

If these builds aren't too slow or cause too much contention, we can
enable them in more builds. Possibly even include them in all
builds that run integration tests. We'll see.

Fixes: #4228