feat(storage/benchmarks): measure warm downloads (#4299)
Download the same file more than once to get data about warm downloads.
Shorten some field names in the output because the plots were unreadable
with longer names.

Not that it matters, because they are fixed numbers and unused in the
plots, but the download vs. upload buffer sizes were reported backwards,
ooops.