feat(storage/benchmarks): quantize app buffers (#4307)
The benchmark uses randomly sized application-level buffers to upload
and download objects. There are some sizes that are known to be much
better. For example, for uploads, anything that is a multiple of 256KiB.
Sometimes it is interesting to test with "bad" buffer sizes, but most
often it is better to use the right values.

There is also no reason to upload and download with the same buffer
sizes, this change uses different buffer sizes for each.

Furthermore, the benchmark called application-level buffers "chunks",
this was a poorly chosen name. There are other "chunks" in a resumable
upload, and this made things confusing.

Finally, changed the default values for these application level buffers
to something that is more reasonable.