ci: patch Abseil base/options.h to pin C++11 types (#4311)
Abseil has a configuration file that controls whether it uses some types
(`std::any`, `std::variant`, `std::string_view`, etc.) from the standard
library, provides its own, or dynamically discovers if they are
available in the library.

If you want to compile Abseil once, and then use with both C++11 and
C++17 compilers you need to edit `base/options.h` and pin these types to
the `absl::` versions. This is what we do in our CI builds, so we need
to patch the file.