feat(storage/benchmarks): control CRC32C/MD5 options
Sometimes we want to run a benchmark with fixed values for CRC32C and/or
MD5 hashing. This adds command line options to run the benchmarks with
either these options fixed to some value from the command-line, or to
chose a random value on each iteration (the default, as before).