ci: remove ci.yml workflow
The goal of adding this workflow was to get some experience w/ GitHub
Actions CI workflows. I think we have that now. Additionally, there's a
fair bit of noise from this workflow as we often see errors from some
tests that don't seem to be legit failures. For example:
https://github.com/googleapis/google-cloud-cpp/pull/4353/checks?check_run_id=743560125

Let's disable this. We can easily add it back later if we want to.