ci: enable su to root in docker containers (#4358)
Fixes: https://github.com/googleapis/google-cloud-cpp/issues/4272

The reason for this PR is that often when we're debugging an issue it
may be easiest to reproduce in the docker container, and to debug it, we
may need to install some debugging tools. We need root access in order
to install this software, and this change gives us that ability. For
example:

```sh
bash-4.2$ su -
Password:
[root@b54ad8707594 ~]#
```

Note that I did not use sudo because sudo requires the invoking user's
UID to appear in /etc/passwd, but since we run our container with
effectively `--user=$(id -u):$(id -g)`, the UID does not appear in the
container's /etc/passwd file and so sudo doesn't work.