ci(macos): less flaky bigtable examples (#4369)
The default for the Cloud Bigtable library is to only retry idempotent
operations. On the macOS builds we are getting a lot of transient errors,
and with the default policy these result in broken builds.

I changed the default policy in the examples, taking care to not make
this part of any example showing how to initialize the library.

I also changed some examples to use safer (idempotent) mutations, because
we should show how to use the service correctly.