fix: docker build on ubuntu (#4380)
On Ubuntu chpasswd fails if the password is the empty string. So we
can't use that trick to enable passwordless `su` on Ubuntu. We should be
able to also make this change by tweaking `/etc/pam.d/su-l`, and this
works on all distros that I tested, including Ubuntu 19+, but it fails
on Ubuntu 18 and 16. So I'm not sure how to enable passwordless `su` on
these Ubuntu distros.

Fixes the issue that @mr-salty mentioned in https://github.com/googleapis/google-cloud-cpp/commit/80b9f3fa6415aed4e75909945a37d2538c3cdd15#commitcomment-39798302