refactor: use absl::optional (#4410)
This PR changes google::cloud::optional<T> to be an alias template for
absl::optional<T>. This is an ABI change, but should be API
compatible.

I also moved the internal::OptionalRow alias into the private section
of the one class where it's used. The TODO referenced by this old alias
had already been closed
(#277).