fix(storage): accept more than one matching log line (#4445)
Some of the integration tests for storage look for specific log lines to
verify the right data is sent to the server. Typically only one such
line is present, but if the library retries two or more might appear. I
ran into this in one of my builds, but seems to be fairly rare.