fix: CompletionQueue::RunAsync is always async
Prior to this change we sometimes ran the callable passed to
`CompletionQueue::RunAsync()` in the calling thread and not
in the thread pool, which was hard to reason about.

As part of these changes I also fixed the code to support
move-only callables, and added a test for it.

The code is "future proofed" to support a `void()` callable.
Which is interesting for a number of reasons (think "executors").