test: reduce flakiness in CreateBucket/DeleteBucket (#4463)
A single project (and we use a single project in our CI builds) can make
at most one CreateBucket or DeleteBucket request every 2 seconds. While
we have taken measures to minimize the number of times we touch
production (e.g. the CMake builds use the testbench, the Bazel builds
cache tests), sometimes there are enough CI builds hitting production
that we exceed this limit.

In this PR we take some additional measures to reduce the problems:

- The ci/kokoro/docker builds based on Bazel now use the GCS emulator
  (aka testbench) for most tests (a few tests must use production, but
  they do not create buckets AFAIK). This leaves the macOS and Windows
  builds running against production.
- Any loops that create buckets with different configurations now pause
  before each iteration. I do not think this slows down the builds
  because those "fast" calls failed and needed to retry anyway.
- For integration tests that create and/or delete buckets we use a
  longer initial backoff (5s) and longer total retry period (10m), as
  the default (for integration tests) were too short to actually work.
- In the examples, where we often create a temporary bucket, we pause
  before deleting the bucket if the total time is less than 2 seconds.
  Otherwise the delete operation will need to retry anyway.