test: deflake background_threads_impl_test (#4470)
Before this PR the tests would (a) pass without problems when not under
load, (b) when under load (a second build running simultaneously) it
would fail about 80 times out of 10,000. After this change the test
passed 20,000 times under load and another 20,000 times without load.
The worst case running time is about 11 seconds, which suggests to me
the shorter 500ms timeouts are not a good tradeoff as they cost some
flakiness in our builds.