ci: fix storage testbench on ubuntu-based builds (#4472)
The Ubuntu-based builds cannot start the storage testbench. This
seems to be a recent problem, as this worked on my workstation
with a cached Docker image, but failed when building the Docker
image from scratch. There are only a handful on Ubuntu-based
builds, and only `integration` and `integration-nightly` actually
use the testbench.

Added some additional logging to diagnose this problem, left it on
because it might help in the future.