test: deflake CompletionQueueTest.ShutdownWithPending
Very rarely this test would fail because the 20ms timer expired before
the very next line could run (scheduling is hard). Fixed the test by (a)
making the timer 500ms, and (b) put a loop to create a new timer if the
first few expired.

In a few cases this would be testing "nothing" as the timer could expire
before the completion queue really is shutdown, but that (rare) false
positive is better (I think) than rare false negatives.