chore: update bazel from 2.0.0 -> 3.3.1
We have several flakes in our CI builds that are related to bazel and
timeouts. These are not really flakes in our code, so it's difficult to
see how to debug or fix them. So as a first step, let's update to the
newest bazel to see if the flakes still exist.

Related: https://github.com/googleapis/google-cloud-cpp/issues/3645
Related: https://github.com/googleapis/google-cloud-cpp/issues/4045