docs: added question.md issue template (#4486)
Added a "Contact us" section to the main README.md file to encourage users to use the GitHub issues for contacting us. I also added a "Question" Issue template to make it easier and more acceptable for users to contact us with normal questions, kinda like they might uses a mailing list for.