feat(storage): reduce copies during uploads (#4489)
Reduce the number of copies during uploads, change `UploadChunkRequest`
to receive a vector of buffers for the payload, and use that vector when
making the PUT request. We still have copies in the `std::ostream`
layer, those will be fixed in a future PR.

Because most of the time we only send one buffer, I changed CurlRequest
to avoid using read callbacks if possible.