refactor(GCS+gRPC): simplify resumable upload code (#4488)
The streaming uploads have a number of restrictions on message sizes,
chunk quantums, and final message flags. This refactors the code to keep
all this complexity in one place. It also dtakes advantage of the
`google.storage.v1.ServiceConstants` proto which did not exist or I
missed when we originally implemented this code. The code has more
conditional code now, but does not create more data copies than before.
Unfortunately a copy is more or less unavoidable because protos need to
own the payload and we cannot move it (both because we might need it on
a retry and because it is received as a `std::string const&`).