feat(storage): reduce data copies in uploads
With this change it is possible to avoid all copies of upload data until
the libcurl/openssl layer (at which point a copy is unavoidable). We say
"possible" because there are cases where some amount of data copies are
necessary (more on these later).

Some background first: resumable uploads in GCS work by uploading
"chunks". The size of these chunks (except the last one) must be
multiples of the upload quantum (256KiB). Uploading a chunk that is not
a multiple of the quantum implicitly finalizes the upload.

This means that any data smaller than a quantum *must* be buffered.
Applications should consider always providing buffers that are a
multiple of the quantum to avoid such buffering.

With that said, many applications upload data by calling
`std::ostream::write()` repeatedly, as long as the buffer provided in
these calls is larger than `clientOptions::upload_buffer_size()` the
buffer will be passed without additional copies to libcurl. Any portion
that is not a multiple of the quantum is not uploaded and buffered.

As an additional improvement this change fixes `sync()` to actually
flush data if the buffer contains at least one multiple of the upload
quantum (256KiB). Smaller buffers cannot be flushed because GCS
finalizes an upload on the first chunk that is not a multiple of the
quantum.

The cases where copying are unavoidable include:

- Data provided via operator<<() is always buffered before copying.
- Buffers smaller than the GCS upload quantum need to be copied.
- Buffers that are not a multiple of the upload quantum are partially
  copied.

Also fixed the testbench to work with chunked (these are HTTP chunks,
not GCS chunks) transfer encoding for resumable uploads.