fix: use flag to prevent vcpkg lock timeout
Fixes https://github.com/googleapis/google-cloud-cpp/issues/4487

The flag was implemented by Microsoft in
https://github.com/microsoft/vcpkg/pull/12227, thanks to @strega-nil!

I believe we were seeing this flake not because of multiple installs
happening at once, but rather because the machines that run our docker
tests were sometimes overloaded and running very slow and thus it
sometimes took more than 1.5 sec to acquire the lock. This is just a
hypothesis.

But regardless, we're now passing `--x-wait-for-lock` to all of our
`vcpkg install` invocations to avoid these flakes.