refactor(storage): upload chunks can have many buffers (#4491)
This change allows the `ObjectWriteStreambuf` to only upload a portion
of a larger buffer via a `ResumableUploadSession`, that removes one copy
from the upload path.

In addition, `ResumableUploadSession` now accepts more than one buffer
as the payload, which is a necessary step to remove more additional data
copies from the upload path.

We also use this in the retry layer to avoid a copy when only part of the
data was committed, as the retry can just shrink the buffers without copying
them.  This copy was rare, but nice to remove it too.