fix: avoid unnecessary chunk to finalize uploads
When closing a stream we need to send a chunk to finalize the upload. If
there was more than one upload quantum in the put area we first flushed
that data with one chunk, and then sent the finalization chunk, with
many zero bytes. It is more efficient to avoid the flush and send a
single finalization chunk.