ci: deflake macOS cache updates (#4505)
The last few failures for #3645 the build timed out because
uploading or downloading the cache made no progress in
3600s.  With this change the upload and/or download of the
cache may timeout (in about 1200s), but that would not break
the build because they run with a `|| true` at the end.

Whether the overall build will still timeout for other reason, or
the 1200s consumed by the cache upload/download timeout
proves too much, that remains to be seen. In any case, I believe
this would help, if nothing else, it would give us more
information about what failed.