fix(storage): avoid chunked transfer encoding
In some (we think relatively rare) cases we were using chunked
transfer encoding, which wastes bandwidth when the size of the transfer
is known (as it is in our case).

The problem only manifested itself when an upload stream had some data
in its buffers and the application provided a large enough call in
`.write()` to fill an upload chunk.

This change adds some metadata in the testbench to detect the transfer
encodings used for an upload, that seems better than examining the log
of the upload.