feat: support moving a Transaction into an error state
Change the signature of the transaction `Visit()` functor to take a
`StatusOr<TransactionSelector>&` instead of a `TransactionSelector&`.

This allows a functor that failed to allocate a transaction ID to
move the transaction into an error state (with a `Status`) instead
of leaving it as a "begin", for the use of future visitors like
`Commit()`.

Only `ConnectionImpl::CommitImpl()` and `RollbackImpl()` have been
modified to check for the error state.  Every other `ConnectionImpl`
transaction visitor currently assumes there is no error (which is OK
as no one currently imposes one).

Part of #4516.