feat(pubsub): a skeleton for PublisherConnection (#4552)
Implements a skeleton for `pubsub::PublisherConnection()` with a single
`Publish()` member function. The function returns a
`future<StatusOr<std::string>>` because eventually needs to be
implemented by an asynchronous RPC, but for now we just return an
immediately satisfied future and use the blocking RPC.

Introduce a `pubsub_testing::MockPublisherStub` so we can write unit
tests for this new class. Also wrote the first unit test.