feat(pubsub): a skeleton for the Publisher client
This PR introduces a very naive version of the `pubsub::Publisher`
client. It also introduces a (somewhat trivial) unit test for the new
class, and a mock for `pubsub::PublisherConnection`. It also introduces
the configuration options for batching and ordering keys, though these
are not used (yet).

This is all going to be more interesting once we implement some amount
of batching.