feat(pubsub): a skeleton for SubscriberConnection (#4558)
Implement an initial version for `pubsub::SubscriberConnection`, leaving
a number of TODO entries for (near) future improvements. Add the
necessary mocks to write unit tests, to do so this change also
introduces the `pubsub_client_mocks` library with user-facing mocks (as
opposed to the mocks that we only expect we would need in our own
tests). Made sure the tests are all passing.