feat: store a Status when invalidating Transactions
after further discussion, @devbww and I reversed the decision in #4545
to use `optional` to mark a `Transaction` invalid, and instead we will
store and propagate the `Status` of the failed `BeginTransaction`.

we also discussed how `Rollback` should function in various failure
scenarios, and I did some testing making manual RPCs to spanner.
When `Rollback` is called with an invalid `Transaction` it now
propagates the `Status` instead of returning `OK`. I also left a TODO
to change the behavior in the `has_begin` case to better match backend
behavior.

part of #4516