chore: remove no-longer-needed docker mount (#4679)
It looks like since our check-style.sh script now uses git ls-files
we probably don't need to mount the cmake-build-debug/ directory, which
CLion uses by default. The only downside is that this directory is
created in everyone's repo dir (even if they don't use clion) as a dir
owned by root. So if we don't need it, and our CLion users' workflows
still work, it seems good to get rid of.