ci: reduce root disk usage on Windows (#4692)
We are informed that the problems in #4363 might be caused
by the CI machine running out of space in the boot partition.
This change aims to reduce the chances of that happening,
and to improve our ability to troubleshoot for that case.

Remove temporary files after they are no longer used, for
example, when refreshing the `vcpkg` cache we may need to
create all the `.o` files for the dependencies, but once these
are installed the `.o` files are no longer needed. Likewise, the
tarball for the Bazel cache is not needed after it is extracted.

Add some logging to report how much disk space is left before
and after some of the steps that consume significant disk.