fix: explicitly begin failed implicit begin transactions
If an operation which has `TransactionSelector::begin` set in order to
implicitly begin a transaction fails, explicitly call `BeginTransaction`.

If that fails, invalidate the `Transaction` so all subsequent operations
fail - most importantly `Commit`, which is necessary to maintain atomicity
since the first operation did not execute.

Otherwise, retry the original operation using the transaction ID
obtained from `BeginTransaction`. That may fail or succeed, but even if
it fails, it may be a "non-constraint error" (read errors and certain
types of write/dml errors) which permit the rest of the `Transaction` to
succeed. Beginning the transaction leaves it to the backend to make
that decision.

Fixes #4516