ci: parallelize check-abi build step
The "check-abi" part of the build was done sequentially taking ~7
minutes on my machine. After this change it takes about 2.5 minutes.

I also changed the "update-abi" build to only update the expected ABI
files and skip the compliance check. The rationale for this change is
because if the user is asking to update the expected ABI is probably
because a previous "check-abi" run failed and that's how they know they
need to update the ABI.

One final small change is that the names in the report are changed from
"reference" and "current" to "expected" and "actual", which matches the
file names better.