Treat an "OK" response with no transaction as invalid.
If we requested a transaction with `TransactionSelector::begin` in a
request but the response does not contain a `Transaction`, that is
an invalid response and we now consistently treat it as such by marking
the transaction as invalid and returning an error. I added a test for
that.

also, if an operation fails, and the subsequent `BeginTransaction` also
fails, return the operation status rather than the `BeginTransaction`
status.