fix: explicitly begin failed implicit begin transactions (#4706)
If an operation which has `TransactionSelector::begin` set in order to
implicitly begin a transaction fails, explicitly call `BeginTransaction`.

If the `BeginTransaction` fails, invalidate the `Transaction` so all
subsequent operations fail - most importantly `Commit`, which is necessary
to maintain atomicity since the first operation did not execute. Note the
returned status from the initial operation is that operation's status rather
than the `BeginTransaction` status, but all subsequent operations will fail
with the `BeginTransaction` status.

If we successfully obtained a transaction ID from `BeginTransaction`, retry
the original operation using that ID. That may fail or succeed, but even if
it fails, it may be a "non-constraint error" (read errors and certain types
 of write/dml errors) which permit the rest of the `Transaction` to succeed.
Beginning the transaction leaves it up to the backend to make that decision.

If we set `TransactionSelector::begin` in a request but the response does
not contain a `Transaction`, the response is invalid and we now consistently
treat it as such by marking the transaction as invalid and returning an error.

Fixes #4516