ci: workaround Windows+CMake+x86 build problems (#4734)
Our builds are failing because one executable is reportedly created by
the CMake step, but the executable is no longer present by the time we
get to the CTest step.  Creating the executable first with CMake, and
then building everything else *seems* to workaround whatever the problem
is. This is a obviously brittle, but the best answer we have at the
moment.