feat: log the result for future<StatusOr<T>> (#4746)
When a function returns `future<StatusOr<T>>` we cannot log the results
immediately, they may not even exist! Previously we just logged whether
the returned future was ready, now we attach a callback to the future
which logs the value whenever it becomes available.

Because this may be much later in the log, the application might need a
token to match the request and response pairs. I am using a per-thread
counter and the thread id as an identifier.