feat(pubsub): expose delivery_attempt field (#4807)
When using dead-letter queues (DLQ, an upcoming feature for Cloud
Pub/Sub) the message might include a "delivery attempt" counter, which
approximately tell us how many times has Pub/Sub tried to deliver a
message. We need to expose this field, and choose to expose it as part
of the `AckHandler`, like we do for `ack_id()`.