feat(pubsub): bootstrap snashot CRUD
Bootstrap the snapshot CRUD functions. I cannot start with just one
function, that is not really testable, and I really don't want to create
a giant PR with all of them. This is basically a minimal set to enable
an integration test.

To keep this PR small we do not add the new functions to
`pubsub::SubscriptionAdminClient`, nor do we add the examples for them.