feat(pubsub): implement UpdateSnapshot() (#4846)
Implemented the function in the *Stub, *Connection, and *Client classes.
Updated the unit tests, integration tests, and samples to include this
new function. I had to disable some of the tests when running against
the emulator because the emulator does not support this function. Filed
a bug and updated our tracker for emulator workarounds.