feat(pubsub): orderly shutdown subscription sessions
So far it has been easy to shutdown subscriuption sessions as the number
of operations outstanding is 1 or 2 (including the timers). As we
increase the concurrency we will need better code to handle potentially
multiple callbacks in flight, and maybe even multiple calls to
AsynPull() in flight. We also want to nack any pending messages that
will not be delivered to a callback as early as possible.

This change creates a helper class to track all asynchronous operations.
Recall that subscription sessions are just represented by a
`future<Status>`. When this future is canceled that stars the orderly
shutdown process. First, a helper class stops scheduling any additional
work. Next, the timer for the session is canceled. A helper class tracks
how many operations are pending, when the counter reaches 0 the future
is satisfied.

Once the shutdown process is started, any responses to `AsyncPull()`
create a fire+forget asynchronous function to nack the messages received
in that pull request. Should that fire+forget operation fail then Cloud
Pub/Sub will timeout the lease for these messages anyway, so they would
be nacked later, but no data is at risk.