doc(pubsub): more sample aliases (#4866)
Depending on the language the `pubsub_quickstart_subscriber` and
`pubsub_quickstart_create_topic` examples are either full programs, or
incomplete programs with "put your topic ID here", or aliases to
examples in a larger program. I chose the latter for C++ because at
least that makes sure we always compile and test these samples.