ci: enable remote build cache for bazel
This tells all of our bazel builds to use a remote cache in a GCS
bucket. This will apply to all of our *docker* bazel builds, including
presubmit and CI builds. This should make both builds a bit faster, by
shifting CPU work (compiles) to network traffic (fetching cached
artifacts). This is probably a good trade off for our Kokoro machines.

Note: This PR is only changing thigns for our docker builds so we can
get experience with this `--remote_cache` option on bazel. The
ultimately plan would be to eneable this for macOS and Windows as well.

Related to: https://github.com/googleapis/google-cloud-cpp/issues/4783