feat: add a `KmsKeyName` class
I pulled this out of my CMEK branch since it was easily separable and
reduces the size of the diff there. I put it in "common" because we may also
want to use it for storage/ etc.

@devjgm and I already discussed this a bit on #4468 - note I decided to bring
back the single-argument constructor we had discussed eliminating - it's
useful for constructing the object from the string in the proto, and I
used it in the sample code I wrote.

part of #4312