feat(storage): custom_time in ObjectMetadata
Support the `customTime` field in `gcs::ObjectMetadata`. This is a
timestamp field, so it is represented by
`std::chrono::system_clock::timestamp` and the name in C++ is
`custom_time`. Because this field can be changed, we need support for it
in `Patch` and `Update`. I also wrote an example.