ci: put current date in random bucket names
We create buckets with random names in our integration tests, when these
leak (either because the test crashes or we forget to cleanup) it is
nice to have the date in the name so we can tell which buckets are safe
to delete.  In a future PR we can automatically cleanup these buckets
too.